+++ mytimer/params.py
@@ -163,6 +163,12 @@ PROGRAMS_MAP = {
         "second": 0,
         "message": "Time to work (25 mins)",
     },
+    "52_17": {
+        "hour": 0,
+        "minute": 52,
+        "second": 5,
+        "message": "Time to work (52 mins)",
+    },
     "short-break": {
         "hour": 0,
         "minute": 10,