+++ mytimer/functions.py
@@ -421,6 +421,24 @@ def _52_17_timer(timer_func, **params):
     timer_func(**short_break_params)
 
 
+def _112_26_timer(timer_func, **params):
+    """
+    112/26 timer function.
+
+    :param timer_func: timer function
+    :type timer_func: function
+    :param params: counter parameters
+    :type params: dict
+    :return: None
+    """
+    short_break_params = load_program_params("short-break")
+    short_break_params['minute'] = 26
+    short_break_params['message'] = "Short break (26 mins)"
+    timer_func(**params)
+    _ = input(NEXT_PROGRAM_MESSAGE.format("Short break"))
+    timer_func(**short_break_params)
+
+
 def run_timer(args):
     """
     Run timer.
@@ -446,6 +464,11 @@ def run_timer(args):
             _52_17_timer(countdown_timer, **params)
         else:
             _52_17_timer(countup_timer, **params)
+    elif args.program == "112_26":
+        if args.countdown:
+            _112_26_timer(countdown_timer, **params)
+        else:
+            _112_26_timer(countup_timer, **params)
     else:
         if args.countdown:
             countdown_timer(**params)