+++ mytimer/params.py
@@ -169,6 +169,12 @@ PROGRAMS_MAP = {
         "second": 0,
         "message": "Time to work (52 mins)",
     },
+    "112_26": {
+        "hour": 1,
+        "minute": 52,
+        "second": 0,
+        "message": "Time to work (112 mins)",
+    },
     "short-break": {
         "hour": 0,
         "minute": 10,