+++ mytimer/params.py
@@ -73,7 +73,8 @@ PROGRAMS_DEFAULTS = {
 }
 
 BREAKS_DEFAULTS = {
-    "alarm": 1
+    "alarm": 1,
+    "tone": 2
 }
 
 PROGRAMS_MAP = {
@@ -223,5 +224,17 @@ BREAKS_MAP = {
         "minute": 20,
         "second": 0,
         "message": "Animedoro break (20 mins)",
+    },
+    "pomodoro-short-break": {
+        "hour": 0,
+        "minute": 5,
+        "second": 0,
+        "message": "Pomodoro short break (5 mins)",
+    },
+    "pomodoro-long-break": {
+        "hour": 0,
+        "minute": 15,
+        "second": 0,
+        "message": "Pomodoro long break (15 mins)",
     },
 }