+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- 2 new programs
+	1. `52-17`
+	2. `112-26`
 ## [1.1] - 2023-12-20
 ### Added
 - 1 new program