+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `long-break` program duration changed from `30 minutes` to `15 minutes`
 - `pomodoro` program updated
 - `play_sound` replaced by `nava.play`
+- Python 3.5 dropped
 ## [1.2] - 2024-02-05
 ### Added
 - `feature_request.yml` template