+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - 1 new program
 	1. `animedoro`
+- `SECURITY.md`
 ### Changed
 - `nava` added to `requirements.txt`
 - Sound playing system updated
@@ -15,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `short-break` program duration changed from `10 minutes` to `5 minutes`
 - `long-break` program duration changed from `30 minutes` to `15 minutes`
 - `pomodoro` program updated
+- `README.md` updated
 ## [1.2] - 2024-02-05
 ### Added
 - `feature_request.yml` template