+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - 1 new program
 	1. `animedoro`
 - `two_step_timer` function
+- `print_message` function
 - `SECURITY.md`
 ### Changed
 - `nava` added to `requirements.txt`
@@ -17,6 +18,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `short-break` program duration changed from `10 minutes` to `5 minutes`
 - `long-break` program duration changed from `30 minutes` to `15 minutes`
 - `pomodoro` program updated
+- `pomodoro_timer` function modified
+- `load_params` function modified
 - `README.md` updated
 ### Removed
 - `animedoro_timer` function